libswan: add auth keylen check in parser_alg_info_add()
It's the wrong place to do this at the parser level (because we need
those values too for incoming key sizes, so we just end up duplicating
this). It should also use a table not a switch. But it was added here
for now because that's where we have the enc key lengths checked as well.

this resolves testing/pluto/algo-pluto-07