testing: change core dump method in swan-prep
We were not successfull when setting the /proc/sys/kernel/core_pattern
pattern to file pattern. It seems to just break dumping core completely.

Using a pattern starting with "|" does work. Since the core comes in
on stdin, and we cannot use > redirection like a shell, we hack it
using the tee command.