ipsec: Suppress ElectricFence banner via ipsec cmd
Set EF_DISABLE_BANNER=1 in the ipsec command.

The banner contains some annoying blanc lines that are hard to filter out,
causing tests with or without electric fence to be different. To get rid
of the blanc lines, we have to suppress the entire banner including the
copyright message.

Various attempts to coerce systemd into accepting this environment
variable via EnvironmentFile= or eval options in the systemd service file
or even via the systemctl set-environment command failed to achieve this
without modification of the ipsec command.

Apologies to Bruce Perens for suppressing his copyright message for
everyone :P