pluto: ignore and warn about empty passphrase for NSS private keys
README.nss used to recommend adding an empty passphrase for NSS keys in
ipsec.secrets but commit 96d89e16 made any extra tokens errors which breaks
old ipsec.secrets files when used with new libreswan versions.

Ignore an empty passphrase token for now and log a warning about it saying
it'll be an error in a future release.

this is rhbz#1144941 (fedora) and rhbz#1145231 (rhel7)