pluto: Always send modecfg banner and domain if they are set
According to the documentation they ought to be sent when the operator
configured them. Likewise, the split tunelling information should be sent
if the network is configured.

We loose track of whether they have been requested anyway thus we wouldn't
send them at all. Certain clients require them though (libreswan won't
invoke the NetworkManager helper if DNS info and DOMAIN is not set.)