IKEv2: do not pad IKE messages to a multiple of 4 octets
In IKEv2, there is no padding specified in the RFC and some
implementations will reject IKEv2 messages that are padded. As there
are no known IKEv2 clients that REQUIRE padding, this commit changes
the padding code to never perform padding for IKEv2.

The man page for the ikepad= option was updated to reflect this.

The option has not been renamed to be ikev1 specific - I'm a little worried
we will eventually find IKEv2 implementations that require padding :/