* pluto/rsasigkey: Allow specifying seedbits for NSS RNG
* Added --seedbits <numbits> option to pluto and rsasigkey
* Added seedbits= to ipsec.conf "config setup" section

There is a BSI requirement that we seed at least 440bits of random
from /dev/random into the NSS via PK11_RandomUpdate() before we
are allowed to pull random from PK11_GenerateRandom() despite the
fact that NSS already deals on it own with initialising its PRNG.

Since this can seriously stall startup on low-entropy machines,
we do not inflict this upon everyone.

rsasigkey already fed 480 bits into NSS, so I turned this into
an option with 480 as default. pluto's default is to not do this.

A big warning was added to the code using /dev/random in plutomain.c
to ensure people don't change this in the future and to ensure the
function to read from /dev/random is never exported to other parts of
the code.