remove doc-base entry for libreswan-doc
the old libreswan-doc package referred to files that don't exist any
longer.  docs/ipsec.html exists, and we could have pointed to that,
but it appears to be out of date (it references a non-existent
glossary.html talks about FreeS/WAN), so i don't think we want to
point to it specifically.