pluto: Use "third best" monotime() alternative on mismatched kernel/glibc
We had a user report hitting the passert(r ==0). This can happen when the
glibc-headers used to compile libreswan assumes support for a clock that
is not present in the kernel.

Since we already had an alternative for when that clock is not there,
use the alternative instead of passerting.