pluto: NFLOG support for all and per-connection IPsec traffic
- New config setup option nflog-all=XX
- New per-conn option nflog=XX

The config setup nflog-all option uses iptables to run all pre-encrypt
and post-decrypt traffic through the NFLOG target of the specific nflog
group number.

the per-conn nflog= option does the same but only applies to the traffic
covered by the connection. Different connections can use different nflog
group numbers.

traffic can be displayed using a variety of tools, such as wireshark, dumpcap
and tcpdump. For example when setting nflog-all=50, one can issue the following
command to see all traffic before encryption and after decryption:

	tcpdump -n -i nflog:50

(note that traffic filters on the command line are not supported at this time in
 tcpdump)

The command "ipsec sniff" is a shorthand for running tcpdump on the nflog-all
group number.

A group number of 0 means disabled.

If an external firewall management tool such as firewalld or shorewall is used,
these settings might interfere and/or become undone during operation.