pluto: don't run test vectors when in hobbled FIPS mode
we are in fips mode when fips=1 and /etc/system-fips exists. When
/etc/system-fips does not exist (dracut-fips removed) then we
are not in fips mode. But NSS disagrees and does not let us run
our test vector code because it uses functions not available in
fips mode.

So for now do not run test vectors when either of the two conditions
is true.