Update ipsec.conf.in
Without the "exec", you have a shell hanging around, and upstart thinks the shell's PID is the main process. This may confuse the pre-stop part. If you "exec" pluto, then it takes over the PID of the shell that starts it, and upstart monitors the correct PID. Then, when it is shut down in the pre-stop, upstart should not try to restart it. Let me know if it works.