pluto: ensure we cannot get "attempt 10 of 4"
Various combinations of timed events can lead to this. The check was for
try_limit != try which I changed to try <= try_limit

It still logs one more "attempt" than max, because the log line is printed
too early when the decision to abort or continue hasn't been made yet.