_updown.netkey: fixup nflog for host-host and net-net tunnels
When a host-host or net-net connection comes up, we call
route-host/route-client, which is the same target, so we have one addnflog
entry. But when these go down, we call down-host and down-client, which
are two different targets in the updown script.

This was found via the nflog-03-conns test case

(TODO: add for v6 targets too)