testing: added  bunch of logging calls for timed events
changed the passert() in event_schedule(). It checked for time_t > 0
but that is weird as time_t is not a signed value. Changed it to a
pexect for 1 month into the future. pexect will cause warnings for
our test cases, but won't cause crashes if people due use unexpectedly
long values of over a month.