testing: removed obsolete interop-ikev2-strongswan-01-noconn
I converted it to kvm but turns out it is a really silly test from when
our ikev2 code was shiny new, and just tests we didn't crash on an R1
NOTIFY packet. It only used strongswan because openswan at the time did
nto support the NOTIFY payload for ikev2. So, I deleted the entire test.