PROPOSED change: find_sender()'s assumption on pacet length is wrong
It tried to only when the sender was known, but find_sender()
checks if our transmitted packet length is equal to the received
ICMP data. But this length does not match at all, the ICMP is much
shorter. However, it seems the memeq() of the packet_len does equal,
so try to use that instead.

When running with logging in this function, it showed that about 5 out 6
packets would find a sender with this change. The 6th one would still
have a sender == NULL, so this would still cause quite the DOS. See next
commit for a proposed workaround on that.