PROPOSED: check_msg_errqueue() should not log ICMP messages for unknown senders
Running on a busy AUTH_NULL server we get several packets per second back that
we cannot seem to match to a sender, and these were getting logged. So I
changed it to log ICMPs only if DBG_OPPO or we have a sender that does not
have authnull. no known sender means only log with DBG_OPPO

With this and the two previous commits, the nssec.nohats.ca server produces
very little logging.