pluto: ikev2 retransmit detection logic was broken
It would false mark informational responses (after the first two) as
duplicates.

It would also flag duplicates when receiving an IKEv2 Request, and
not answer, but that would only lead to receibing a duplicated request.
So it is better to just process the duplicate on the first go and answer.