pluto: support for pluto --impair-force-fips
This options sets kernel and system mode to fips and skips the hmac
check files to allows pass. This allows pluto running in fips mode
without needing the .hmac files or fips=1 on the command line.

Care should be taken that the NSS library needs to be put in FIPS
mode seperately if fips=1 is not used on the kernel command line.

This option is not supported for whack, because the FIPS mode
decision is made at startup before pluto acecpts any whack commands,
and cannot be changed later after boot. (A comment is left in whack.c
so I don't add this later on by mistake :)