pluto: fix "story" so const need not be cast away
In various bits of DH code, a "story" variable or parameter is used.
Although story was declared without the const attribute, it is set to a
const char * value.  To evade diagnostics it was jammed in via a cast.  Fixed.

Also, enum_show was used to generate the value.  The resulting value is stored
in a global static buffer and is thus subject to surprising overwriting.  I switched
to using enum_name which has no such problem.  I assume that the group in use
actually has a name known to us.