IKEv1: Remove all IPsec SA's of a connection when newest SA is removed.
This behaviour is similar to "ipsec auto --down connection-name"

This resolves an interop issue with Cisco where after a brief outage,
sometimes the connection results in two IPsec SA's being established.  In
this case, after sometime, the cisco router sends an ISAKMP Delete/Notify
message to delete one of the IPsec SAs. If the removed IPsec SA is the
first SA, it will be fine. But if the removed IPsec SA is the newest SA,
the IPsec tunnel state is set to "perspective eroute". And now traffic
between the Cisco and libreswan on the ipsec tunnel is blocked.