pluto: Introduce STF_DROP to drop an IKE packet without response
This is used when in IKE_INIT request is received but hits an OE
clear foodgroup. There is no point sending the message as it is
unauthenticated and cannot be trusted by the initiator. And the
responder is revealing itself to the initiator while it is
configured to never talk to that particular initiator. With this
changes, the system does not need to enforce this policy using a
firewall.

Note that this technically violates the IKEv2 specification that
states we MUST answer (with NO_PROPOSAL_CHOSEN)

This commit also removes the obsoleted/unused STF_STOLEN that was
part of the old taproom code.