testing: newoe update to reduce timing false positives
Using ping -w 2 would cause us to sometimes send 2 and sometimes
send 3 pings. Since -w was only needed when we locally had a block
rule for OE, I changed all other test cases to not use -w anymore.

When pinging with -c 2 to trigger OE, the first packet gets eaten
by the ACQUIRE. The second packet would run a race with IKE and
sometimes win and sometimes lose. To prevent this race, send 1 ping
instead. It will always just be eaten. Note that this caused the
trafficstatus output to change since it no longer reports any byte
count as the tunnel remains unused at this point.

The final ping had a count of 1. I increased it to 2 to ensure we
see what we are expecting and not just had a random packet loss of
1 packet.