pluto: remove two bogus state machine entries
"I2: process INFORMATIONAL"
"R1: process INFORMATIONAL"

These states have no authenticated parent sa and so should never even
process an informational. Only R2 and I3 states are allowed to do so.

These were added by mistake because the SEND_V2_NOTIFICATION() related
code could cause an IKE_AUTH or IKE_INIT request to be answered with
and INFORMATIONAL reply, which is incorrect. The code that did so was
fixed a few commits before this commit.