pluto: Added per-conn option fake-strongswan=yes|no (default no)
Send the strongswan (unversioned) Vendor ID when enabled.

This is to allow esp=twofish and esp=serpent with strongswan without
requiring modifications to strongswan. It will not use private number
ranges for algorithms unless it sees the strongswan vendorid.

Ironically, the twofish/serpent algo private numbers used were taken
from freeswan and used in openswan and libreswan as well.