testing: update ikev1-hub-spoke / ikev2-49-hub-spoke to avoid race
north and west would establish their runnel in different order
depending on the test machine. This results in east's output to
be different, since ip xfrm state would list these in order of
installation in the kernel.

Adding a sleep to westrun.sh to ensure north always wins the race
fixes this.