testing: delete-sa-01: add --no-esn to whack to match output
This test did not inherit the default ESN_NO because it does not
use a config file but whacks directly. The reference output was
changed for all tests to have ESN_NO, so that caused this test to
false positive fail.

Since it makes sense to test the whack optio --no-esn, this was
added to the whack line instead of removing +ESN_NO from the
policy bits in the reference output.