testing: removed delete-state-01. The test no longer makes sense.
It tested using whack to delete the ipsec state, then delete the
ike state. But once west sends the delete/notify to for the ipsec
sa to east, east deletes it and sends a notify. Then sees there is
a childless ike sa and deletes it too, sending a notify to west.
Then west processes the delete and deletes the IKE SA, so by the
time westrun.sh reaches the whack for state #1, it is already gone.