pluto: eliminate cert leak in load_nss_cert_secret()
When the cert wasn't RSA, and when the cert's secret was found, the
reference to the cert was being destroyed.

Fixed by passing the cert (and not the nickname) into
load_nss_cert_secret() and then making the callers responsible for
destroying the cert.  This, in turn, allowed the elimination of two
redundant calls to PK11_FindCertFromNickname().