testing: create swanfedorabase.qcow2 directly using virt-install.
Was creating swanfedroabase.img, then cloning it to swanfedroabase.qcow2.

I suspect the motivation was "a raw image file is much faster",
but that is only measurable when the install isn't twiddling its
thumbs waiting for the network.

This also elimnates the hack to see if the kernel barfed while
cloning .img to qcow2.