web: generate the json files by data-mining the original logs and git repo
For instance, instead of results.json containing redundant
pass/fail/unresolved/untested counts, use "jq" to compute the
counts from the individual test data in results.json.

"kvmresults.py" is used to generate the basic data, and "jq"
is used to manipulate it.  The new scripts build-summary.sh
(and summary.sh), and build-results.sh (and results.sh),
which can be run standalone, contain the details.