web: fix commit ordering by using the commit's "rank"
Where "rank" is defined as the number of commits since the repo's
epoc.  That is: git rev-list --count --first-parent <rev>

This fixes both the previous result links and the graph lines
where, due to out-of-order commit dates, things were, well, in the
wrong order.

Note that, because the graph is using the commit-date for the X-axis,
the out-of-order commits also cause the graph lines to form nice
visual knots.