KLIPS: fix for unregister_netdevice: waiting for XXX to become free.
Use netdev_notifier_info_to_dev() for 3.6.11 and up

On 3.6.11+ kernels, KLIPS could not match the stored physical device
pointer with the saved pointer, and would ignore the device_event. This
in turn let to a refcount of non-zero for these devices, causing those
to fail to unload. This would show up mostly for devices that come and
go, such as ppp or bond devices.

Signed-off-by: Paul Wouters <pwouters@redhat.com>