verify: python3 fixup and Opportunistic Encryption check removal
The new OE does not use a keyword and detecting if it is on was
using an obsolete method.

Added check got ERROR and Warning strings to flag obsolete options

Remove check for hardware random. This is all handled within the
kernel itself now and the user cannot do anything wrong