IKEv1: some IKE SA failure on the initiator would lead to hanging state
This initiator getting STF_FAIL in its state would have its event removed,
and would never do anything anymore. It would not retransmit or timeout.
Errors with STF_FATAL were not affected.

As a side-effect, the whack would also not be released and the ipsec
auto --up command would hang.

This could be seen in various test cases, including ikev1-impair-gx-02
and nss-cert-10-notyetvalid-responder as it caused the test to timeout.