pluto: replace v1auth_to_v2integ() with ikev1_get_esp_info_integ_desc() et.al.
And so deleteth one of the many nasty ESP/AH ikev1<->ikev2 switches.

As a result, for IKEv2 IKE/ESP/AH, it is assumed that all
algorithms have a 'struct ike_alg' entry.

The same can not be said of IKEv1.  For instance, RIPE is missing.
This is because IKEv2 has never had support for ESP/AH RIPE
(the case was missing from v1auth_to_v2integ()).

While this solution isn't ideal, it is progress towards the
"correct fix" where the magic numbers in "ike_info" and "esp_info"
are repaced by "encrypt_desc", "prf_desc" and "integ_desc" pointers.