X509: Fix memory leak in certificate handling (lsbz#278)
In get_pluto_gn_from_nss_cert, the call to CERT_GetCertificateNames
allocates memory in cert->arena which would only be freed when
cert->arena is freed.  It seems that either cert->arena is never
freed or at least it's not freed for quite a while.  The data
from CERT_GetCertificateNames is used almost immediately after by
add_cert_san_pubkeys and then is apparently no longer needed.  If a new
arena is used for CERT_GetCertificateNames instead of cert->arena then
it can be freed in add_cert_san_pubkeys after it is no longer needed.

Signed-off-by: Paul Wouters <pwouters@redhat.com>

Note: long term solution is to remove pluto_gn and use the nss fields
directly, as pluto_gn is a relic from pre-NSS times.