pluto: when IKEv1|IKEv2, use the IKEv1 ike= DH defaults
While the choice is arbitrary, this has less of an effect on the test
results (otherwise need to tweak all the "IKEv1" tests and force them
into IKEv1 only mode), and increases the odds of an interop (otherwise
when one end is IKEv1 and the other is IKEv1|IKEv2 things like quick
fail).  .

Correct fix is to make IKEV1|IKEV2 go away.