IKEv2: re-instate ISAKMP_SA_established()
This function is used to releease_connection() if this connection
replaces it. It causes the connection to be unrouted, so we can
install our new (e)route. Without this call, the old connection
owns the (e)route and the new connection will fail with a
"cannot route -- route already in use" error.

e