pluto/libswan: Change most ttoaddr() to ttoaddr_num()
These calls were not really planning on doing a DNS lookup. It is
mostly changing a textual IP to a struct ip_address.

foodgroups did appear to support DNS but since it is never refreshed,
this seems like a bad idea as these are only resolved at startup. So
support for DNS in foodgroups was also removed.