packaging: add dependency on ldns and for fedora unbound-devel version
The ldns code is used for the new DNS parsing code for IPSECKEY

The unbound version in fedora is needed to work around a bug where
the unbound-devel package does not install unbound-event.h. In
debian and RHEL/CentOS, this bug is not yet fixed so we cannot yet
put in minimum package versions. However, for RHEL we will package
and ship fixed versions in download.libreswan.org RHEL repositories,
also because the 1.4.20 version in RHEL6 and RHEL7 is missing the
unbound-event.h support.