packaging: Fix spec file for rpm vs libreswan/make variables
The use of %global USE_DNSSEC 1 to select a choice both in the spec
file and in the libreswan makefile did not work, as 1 != true. This
resulted in USE_DNSSEC not getting enabled in the build.

This separates the two kind of variables more clearly, avoiding these
mistakes in the future.