libipsecconf/pluto: distinguish authby=never from default properly
When introducing leftauthby/rightauthby in 3.19, a bug was introduced
that did not properly recognise authby=ever connections with the
NEVER_NEGOTIATE() macro. This was due to the lack of authentication
via both auth= and leftauthby/rightauthby= being seen as a default,
and a symmetric RSASIG policy is then added to the policy. This
meant authby=never connections ended up with an auth policy and so
NEVER_NEGOTIATE() was false.

To fix this, instead of relying on a lack of RSASIG/PSK policy to
determine authby=never, use a specific policy bit to represent this,
using POLICY_AUTH_NEVER_IX.

While fixing this in whack and the parser, it also appeared that
there was no whack option for authby=null, so --auth-null was added.

NEVER_NEGOTIATE() only checked for absence of POLICY_ENCRYPT and
 POLICY_AUTHENTICATE which also