testing: add alias-01 to TESTLIST and update.
This seems to expose the internal connalias= keyword used for subnetS=
connection instantiation.

the test still fails because of the weird config on west.conf that
adds an alias'ed conn via also=. I am not sure if this is really
supposed to work

note connalias= does not appear in the man page.