XFRM: Use proper alignment for IPv4 AH as per RFC4302 Section 3.3.3.2.1
This requires XFRM_ALIGN4 support, which is present since 2.6.39

Warning: this breaks backwards compatibility with older kernels
with XFRM/NETKEY as it does not seem to support both alignment types.

KLIPS work fine with either alignment.

We decided not to add a backwards compatible option for this at this
moment. If you really need to interop with AH and old kernels, let us
know. The better way forward is to use ESP_NULL instead of AH.