parser: reject unavailable kernel algorithms
... as determined by the kernel info tables

This way unavailable algorithms specified by esp= or ah=
are rejected by the parser; and don't reach either the proposal
or backend code.

There is still the problem of IKEv2's hard-wired default proposals
but hopefully the've been selected based on what the kernel should
support (and if not, the kernel should be able to handle that).