kernel: make the struct integ_desc .integ and struct encrypt_desc .encrypt available to netlink
Add .integ and .encrypt to struct kernel_sa.

Rename the overloaded field .encalg->.compalg.  It was being used for both
ESP and COMP.  The rename makes it clear that it should only be used for COMP
(but ESP uses do remain).

Use &ike_alg_integ_hmac_sha2_256_truncbug for the truncbug.