pluto: don't call sanitize_string() in fmt_log() as it is expensive
fmt_log() is only used to write logs, not to pass anything to
updown. We leave the call in for DBG_log() since if you're in
debugging mode, you're slow anyway.

Strings taken from the network and passed to updown are few, and
those calls go via cisco_stringify() which calls sanitize_string()