pluto: add proper support for disabling leftupdown= execution
You can set either leftupdown="" or leftupdown="%disabled" to disable
forking the updown client.

The whack command remains the same. That is leaving out the --updown
value causes the default updown to run, and using --updown "" or
--updown %disabled will disable the updown from running.

This is nicer then using the hack leftupdown=/bin/true as it prevents
work done to prepare the environment via fmt_common_shell_out()

This is useful on very busy servers that require no updown scripting.