pluto: ikev2parent_inI1outR1() postpone reading NOTIFY payloads
In the first loop over the NOTIFY payloads, only look for the COOKIE
payload and handle it. This might result in no state being created.

We do not need to process NOTIFY's to know whether or not we commit
to a state. Ince commited, read the NOTIFY payloads and directly
process these unto state variables without intermediate variables.

This eliminates one loop over the NOTIFY's to re-find the
v2N_SIGNATURE_HASH_ALGORITHMS notify and eliminates a "Make coverity
happy" passert()