pluto: Use XAUTH_MAX_PASS_LENGTH, not a hardcoded 64
The IKEv1 XAUTH password length is limited to 64 bytes in
programs/pluto/ikev1_auth.c:2106 by

if (st->st_xauth_password.ptr == NULL) {
        char xauth_password[64];

This is a mistake and the constant XAUTH_MAX_PASS_LENGTH (128) shall be
used similarily as MAX_USERNAME_LEN is used for the username.

Therefore I propose a patch to use

if (st->st_xauth_password.ptr == NULL) {
        char xauth_password[XAUTH_MAX_PASS_LENGTH];

so that passwords are not limited to 63 characters + newline (when
propted from stdin). It also improves code consistency.

Signed-off-by: Paul Wouters <pwouters@redhat.com>