pluto: find_phase1_state() should only find matching IKE version states.
This is lsw#302 An IKEv1-only and IKEv2-only conn between the same host can end up sharing an IKE SA

See test case ikev1-ikev2-connswitch-01

Basically, if you have two hosts that are configured for both
an ikev1 only and an ikev2 only connection, when bringing up the
second one, this would be ignored. This would lead to an ikev2=insist
causing an IKEv1 quickmode message.

If the ikev1 uses authby=secret and the ikev2 uses authby=rsasig, then
the second connection would completely fail to establish.