pluto: add support for nat-ikev1-method=none
This will suppress sending the NAT-T payloads, and prevents the peer
from performing NAT-T negotiation. It will thus be forced to ESP and
not do ESPinUDP. This only affects IKEv1. IKEv2 will always send NAT-T
payloads.

This option should not be neccessary, as the peer should be able to
force non-encaps itself if it does not like it, but yet here we are :)