XAUTH: Perform custom state change for XAUTH without ModeCFG
It is not possible to use different smc state machine changes based
on ModeCFG payloads, as the content of these payloads determine
which state change we should do, but in both cases with and without
modecfg, we get some modecfg payload.

What happened until this commit, was that the ISAKMP SA was left
dangling despite also moving on and doing quick mode successfully
in the case of NOT doing modecfg.

But since the dangling state ends up timing out and restarting new
keying attempts, it would delete the established IPsec and IKE SA's.
The test cases xauth-pluto-03 and xauth-pluto-04 suffered from this,
but it was not visible in the test case because it ended before the
dangling state caused damage. The test cases were updated to show
the STATEs so any future regression on this becomes visible.

History showed this was attempted to have been fixed before, but
was partially left in an #if 0 statement.

As most XAUTH also involves ModeCFG, this bad case was apparently
rarely hit.