pluto: fill in complete struct cert_hd - coverity warning
We did not fill in cert_hd.isacert_reserved and cert_hd.isacert_length

I'm not sure why we don't set the length when adding such a payload
for sending. I gues out_struct() calculates the length and puts it
there? Otherwise we would be sending a corrupt payload ?

This is probably still a bug.