config: use lmod_arg() to parse the config line plutodebug=
The good news is that uses the same lookup-table and code as --debug
(a duplicate table deleted); and plutodebug=all,no-xauth should
do-the-right thing.

The bad news is that the config parser has been assimilated by the
borg (i.e., it uses stuff in libswan like lswlog and enum_names).

(a long standing bug where having more than 32 debug classes will
result in truncation is not fixed - we don't yet have 32 debug
classes)