ikev2: responder temp --mobike probe remote-- address to respond
1. Send response of informational exchange, with v2N_UPDATE_SA_ADDRESSES
to the new address.

After responding forget the new remote address.

2. Also IKEv2 livenes, on responder respond to new temp address
when the initiator has a new address, and old adress is removed, it
may send a liveness without MOBIKE payloads. The responder just respond
to the new address. And do not update st_remote_addr (on responder) yet.
It would be updated if the initiator send MOBIKE payaload in the
following exchange.

Initiator use this inforational reply only to
"determine if the current path is still usable". This is not a
MOBIKE exchange. A MOBIKE exchange could follow.

See RFC 4555 #3.5.
Note : this is one of the ways initiator start MOBIKE. There are other
methods which do not involve such a liveness probe.