pluto/lib: Remove obsoleted/unused %myid support
In the 199x OE, there was no NAT and determining our own ID based
on IP address or FQDN was a real option. This isn't true now and
for most of those use cases now we use a null ID. For the other
cases, the FQDN or IP based ID should be set manually if it is static.

If we need to re-gain this to be automatic, it should be handled
only in add_connection()