ikev2: MOBIKE initiator - call routing up-down script
The interface may go down, along with route(s) and and IP address used by
the connection, when switching networks -- WiFi-to-WiFi or switchig to 4G...
When pluto detect the currernt address is removed it start MOBIKE probe.
And call connection "down" and  'unroute'
After MOBIKE succeeds call "up" and "route"

The MOBIKE initiator traffic may leak when switching WiFi to
Cellular or the other way around. Also if there is captive portal,
traffic may leak.

The exisiting sessions such as ssh are reported to survive, short
interruptions.

"down" & "unroute" may also remove the IP address for a brief period.
An improvment to removing the IP and traffic leak could be "mobike-down" and
"mobike-up" verbs in the upddown script.

ikev2: MOBIKE initiator send probe even when the new IP address == old IP address