testing: add --impair send-no-retransmits
This option causes pluto to wait the full retransmit timeout (60s)
while sending no retransmits.  Useful when the other end needs to be
given a little bit more time.

Contrast that with the existing --impair retransmits which only
waits for the initial delay (0.5s).  Useful when a quick timeout is needed.