retransmits: work-around discrepancy between monotime_t and event-loop time
For instance, after retransmits with delays 15s, 15s, and 30s, the
total time should exceed (>=) 60s yet, according to monotime_t, this
is completed in only 59.983s and since that is less than the timeout
(60s here) an extra retransmit gets scheduled.

The work-around is to also accumulate the delays and compare
that against the timeout.

(this problem really becomes apparent when monotime_t has sub second
resolution)