IKEv2: Add support for modecfgdns= and modecfgdomains=
This adds support for INTERNAL_IP4_DNS and INTERNAL_IP6_DNS
and passes these options to updown similar to IKEv1, so that
resolv.conf or the running unbound DNS server are updated.

This adds support for INTERNAL_DNS_DOMAIN as defined by
draft-ietf-ipsecme-split-dns-03

It does not yet add suppor tfor INTERNAL_DNSSEC_TA.

Note this renames the option modecfgdns1= and modecfgdns2= to
modecfgdns= and modecfgdomain= to modecfgdomains=

The new options take a comma or space seperated list.

The old config options are retained as aliases. This does have the
effect that modecfgdns2= overwrites modecfgdns1= as only the last
option is actually stored in modecfgdns=

This also moves the received DNS / DOMAIN / BANNER options from
struct connection to struct state where it belongs.