pluto: flag type asssertions with comments
Some function-type typedefs are useful but cannot be used in function
definitions.

Our rule is: each function definition that cannot use the appropriate
typedef must be within the scope of a declaration of that same
function that does use the typedef.

These type assertions look odd to the casual reader.  They look like
variable declarations.  We add /* type assertion */ to each to help the
reader.

It would be nice to always have such a declaration adjacent to the
definition but GCC whines if there has already been a declaration.  We
used to have such declarations, but as comments.  But comments aren't
checked (or maintained) so I've ditched them.